+++ go/src/github.com/bitrise-core/bitrise-init/steps/steps.go
@@ -37,7 +37,7 @@ const (
 	cocoapodsInstallVersion = "1.5.2"
 
 	xcodeArchiveID      = "xcode-archive"
-	xcodeArchiveVersion = "1.8.2"
+	xcodeArchiveVersion = "1.8.3"
 
 	xcodeTestID      = "xcode-test"
 	xcodeTestVersion = "1.13.7"
@@ -109,7 +109,7 @@ func DeployToBitriseIoStepListItem() bitrise.StepListItemModel {
 // ScriptSteplistItem ...
 func ScriptSteplistItem() bitrise.StepListItemModel {
 	stepIDComposite := stepIDComposite(scriptID, scriptVersion)
-	return stepListItem(stepIDComposite, "Do anything with script step", "", nil)
+	return stepListItem(stepIDComposite, "Do anything with Script step", "", nil)
 }
 
 //------------------------